+++ packages/backend/src/services/ServiceRepository.ts
@@ -799,6 +799,7 @@ export class ServiceRepository
                     dashboardModel: this.models.getDashboardModel(),
                     spaceModel: this.models.getSpaceModel(),
                     schedulerClient: this.clients.getSchedulerClient(),
+                    featureFlagModel: this.models.getFeatureFlagModel(),
                 }),
         );
     }