+++ packages/frontend/src/ee/features/embed/SettingsEmbed/EmbedPreviewChartForm.tsx
@@ -12,15 +12,17 @@ import {
     Flex,
     Group,
     Input,
+    Paper,
+    SegmentedControl,
     Select,
     Stack,
     Switch,
     Text,
     TextInput,
     Title,
-} from '@mantine/core';
+} from '@mantine-8/core';
 import { useForm } from '@mantine/form';
-import { IconLink, IconPlus, IconTrash } from '@tabler/icons-react';
+import { IconEye, IconLink, IconPlus, IconTrash } from '@tabler/icons-react';
 import { useMutation } from '@tanstack/react-query';
 import { useCallback, type FC } from 'react';
 import { v4 as uuidv4 } from 'uuid';
@@ -136,6 +138,18 @@ const EmbedPreviewChartForm: FC<{
         [projectUuid],
     );
 
+    const handlePreview = useCallback(async () => {
+        const state = form.validate();
+        if (state.hasErrors) {
+            return;
+        }
+
+        const data = await createEmbedUrl(
+            convertFormValuesToCreateEmbedJwt(formValues, true),
+        );
+        window.open(data.url, '_blank');
+    }, [formValues, form, convertFormValuesToCreateEmbedJwt, createEmbedUrl]);
+
     const generateUrl = useCallback(async () => {
         const data = await createEmbedUrl(
             convertFormValuesToCreateEmbedJwt(form.values),
@@ -148,124 +162,183 @@ const EmbedPreviewChartForm: FC<{
 
     return (
         <form id="generate-embed-url" onSubmit={handleCopySubmit}>
-            <Stack mb={'md'}>
-                <Stack spacing="xs">
-                    <Title order={5}>Preview</Title>
-                    <Text color="dimmed">
-                        Preview your embed URL and copy it to your clipboard.
-                    </Text>
-                </Stack>
+            <Stack gap="md" mb="md">
                 <Select
                     required
-                    label={'Chart'}
+                    label="Chart"
                     data={charts.map((chart) => ({
                         value: chart.uuid,
                         label: chart.name,
                     }))}
                     placeholder="Select a chart..."
                     searchable
-                    withinPortal
                     {...form.getInputProps('chartUuid')}
                 />
-                <Select
-                    required
-                    label={'Expires in'}
-                    data={['1 hour', '1 day', '1 week', '30 days', '1 year']}
-                    withinPortal
-                    {...form.getInputProps('expiresIn')}
-                />
-                <Input.Wrapper label="User identifier">
-                    <TextInput
-                        size={'xs'}
-                        placeholder="1234"
-                        {...form.getInputProps(`externalId`)}
-                    />
-                </Input.Wrapper>
-                <Input.Wrapper label="User email">
-                    <TextInput
-                        size={'xs'}
-                        placeholder="Type an email to add as intrinsic user attribute"
-                        {...form.getInputProps('email')}
-                    />
-                </Input.Wrapper>
-                <Input.Wrapper label="User attributes">
-                    {form.values.userAttributes.map((item, index) => (
-                        <Group key={item.uuid} mt="xs">
-                            <TextInput
-                                size={'xs'}
-                                placeholder="E.g. user_country"
-                                {...form.getInputProps(
-                                    `userAttributes.${index}.key`,
-                                )}
-                            />
-                            <TextInput
-                                size={'xs'}
-                                placeholder="E.g. US"
-                                {...form.getInputProps(
-                                    `userAttributes.${index}.value`,
-                                )}
-                            />
-                            <ActionIcon
-                                variant="light"
-                                onClick={() =>
-                                    form.removeListItem('userAttributes', index)
-                                }
-                            >
-                                <MantineIcon color="red" icon={IconTrash} />
-                            </ActionIcon>
-                        </Group>
-                    ))}
-                    <Group>
-                        <Button
-                            size="xs"
-                            mr="xxs"
-                            variant="default"
-                            mt="xs"
-                            leftIcon={<MantineIcon icon={IconPlus} />}
-                            onClick={() =>
-                                form.insertListItem('userAttributes', {
-                                    key: '',
-                                    value: '',
-                                    uuid: uuidv4(),
-                                })
-                            }
-                        >
-                            Add attribute
-                        </Button>
-                    </Group>
-                </Input.Wrapper>
 
-                <Switch
-                    {...form.getInputProps(`canExportCsv`)}
-                    labelPosition="left"
-                    label={`Can export CSV`}
-                />
-                <Switch
-                    {...form.getInputProps(`canExportImages`)}
-                    labelPosition="left"
-                    label={`Can export Images`}
-                />
-                <Switch
-                    {...form.getInputProps(`canViewUnderlyingData`)}
-                    labelPosition="left"
-                    label={`Can view underlying data`}
-                />
+                <Stack gap="xs">
+                    <Text size="sm" fw={500}>
+                        Expires in
+                    </Text>
+                    <SegmentedControl
+                        value={form.values.expiresIn}
+                        onChange={(value) =>
+                            form.setFieldValue('expiresIn', value)
+                        }
+                        radius="md"
+                        data={[
+                            { label: '1 hour', value: '1 hour' },
+                            { label: '1 day', value: '1 day' },
+                            { label: '1 week', value: '1 week' },
+                            { label: '30 days', value: '30 days' },
+                            { label: '1 year', value: '1 year' },
+                        ]}
+                        size="xs"
+                    />
+                </Stack>
+
+                <Paper p="md" withBorder>
+                    <Stack gap="md">
+                        <Title order={6}>Identification & Security</Title>
+                        <Stack gap="xs">
+                            <Input.Wrapper label="User identifier">
+                                <TextInput
+                                    size="xs"
+                                    placeholder="1234"
+                                    {...form.getInputProps('externalId')}
+                                />
+                            </Input.Wrapper>
+
+                            <Input.Wrapper label="User email">
+                                <TextInput
+                                    size="xs"
+                                    placeholder="Type an email to add as intrinsic user attribute"
+                                    {...form.getInputProps('email')}
+                                />
+                            </Input.Wrapper>
+
+                            <Input.Wrapper label="User attributes">
+                                <Stack gap="xs" mt="xs">
+                                    {form.values.userAttributes.map(
+                                        (item, index) => (
+                                            <Group
+                                                key={item.uuid}
+                                                gap="xs"
+                                                wrap="nowrap"
+                                            >
+                                                <TextInput
+                                                    size="xs"
+                                                    placeholder="E.g. user_country"
+                                                    style={{ flex: 1 }}
+                                                    {...form.getInputProps(
+                                                        `userAttributes.${index}.key`,
+                                                    )}
+                                                />
+                                                <TextInput
+                                                    size="xs"
+                                                    placeholder="E.g. US"
+                                                    style={{ flex: 1 }}
+                                                    {...form.getInputProps(
+                                                        `userAttributes.${index}.value`,
+                                                    )}
+                                                />
+                                                <ActionIcon
+                                                    variant="light"
+                                                    color="red"
+                                                    onClick={() =>
+                                                        form.removeListItem(
+                                                            'userAttributes',
+                                                            index,
+                                                        )
+                                                    }
+                                                >
+                                                    <MantineIcon
+                                                        icon={IconTrash}
+                                                    />
+                                                </ActionIcon>
+                                            </Group>
+                                        ),
+                                    )}
+                                    <Button
+                                        size="xs"
+                                        variant="default"
+                                        leftSection={
+                                            <MantineIcon icon={IconPlus} />
+                                        }
+                                        onClick={() =>
+                                            form.insertListItem(
+                                                'userAttributes',
+                                                {
+                                                    key: '',
+                                                    value: '',
+                                                    uuid: uuidv4(),
+                                                },
+                                            )
+                                        }
+                                    >
+                                        Add attribute
+                                    </Button>
+                                </Stack>
+                            </Input.Wrapper>
+                        </Stack>
+                    </Stack>
+                </Paper>
+
+                <Paper p="md" withBorder>
+                    <Stack gap="md">
+                        <Title order={6}>Interactivity & Permissions</Title>
+                        <Stack gap="xs">
+                            <Text size="sm" fw={500}>
+                                Users can:
+                            </Text>
+                            <Switch
+                                {...form.getInputProps('canExportCsv', {
+                                    type: 'checkbox',
+                                })}
+                                label="Export CSV"
+                            />
+                            <Switch
+                                {...form.getInputProps('canExportImages', {
+                                    type: 'checkbox',
+                                })}
+                                label="Export Images"
+                            />
+                            <Switch
+                                {...form.getInputProps(
+                                    'canViewUnderlyingData',
+                                    {
+                                        type: 'checkbox',
+                                    },
+                                )}
+                                label="View underlying data"
+                            />
+                        </Stack>
+                    </Stack>
+                </Paper>
+
                 <Flex justify="flex-end" gap="sm">
                     <Button
-                        variant={'outline'}
+                        variant="light"
+                        leftSection={<MantineIcon icon={IconEye} />}
+                        onClick={handlePreview}
+                    >
+                        Preview
+                    </Button>
+                    <Button
+                        variant="default"
                         type="submit"
-                        leftIcon={<MantineIcon icon={IconLink} />}
+                        leftSection={<MantineIcon icon={IconLink} />}
                     >
                         Generate & copy URL
                     </Button>
                 </Flex>
             </Stack>
-            <Stack mb="md">
-                <Stack spacing="xs">
+
+            <Stack gap="md" mb="md">
+                <Stack gap="xs">
                     <Title order={5}>Code snippet</Title>
-                    <Text color="dimmed">
+                    <Text c="dimmed" fz="sm">
                         Copy and paste this code snippet into your application
-                        to generate embed urls.
+                        to generate embed URLs.
                     </Text>
                 </Stack>
                 <EmbedCodeSnippet